+++ tests/ZiggyCreatures.FusionCache.Tests/L1L2BackplaneTests.cs
@@ -1,5 +1,4 @@
 ﻿using Azure.Messaging.ServiceBus;
-using Azure.Messaging.ServiceBus.Administration;
 using FusionCacheTests.Stuff;
 using Microsoft.Extensions.Caching.Distributed;
 using Microsoft.Extensions.Caching.Memory;
@@ -9,7 +8,7 @@ using Xunit;
 using ZiggyCreatures.Caching.Fusion;
 using ZiggyCreatures.Caching.Fusion.Backplane;
 using ZiggyCreatures.Caching.Fusion.Backplane.AzureServiceBus;
-using ZiggyCreatures.Caching.Fusion.Backplane.AzureServiceBus.Helpers;
+using ZiggyCreatures.Caching.Fusion.Backplane.AzureServiceBus.AzureServiceBusWrapper;
 using ZiggyCreatures.Caching.Fusion.Backplane.Memory;
 using ZiggyCreatures.Caching.Fusion.Backplane.StackExchangeRedis;
 using ZiggyCreatures.Caching.Fusion.DangerZone;
@@ -17,12 +16,13 @@ using ZiggyCreatures.Caching.Fusion.DangerZone;
 namespace FusionCacheTests;
 
 public partial class L1L2BackplaneTests
-	: AbstractTests
+	: AbstractTests, IClassFixture<L1L2AzureServiceBusFixture>
 {
-	public L1L2BackplaneTests(ITestOutputHelper output)
+	public L1L2BackplaneTests(ITestOutputHelper output, L1L2AzureServiceBusFixture azureServiceBusFixture)
 		: base(output, "MyCache:")
 	{
-		if (UseRedis)
+		_azureServiceBusFixture = azureServiceBusFixture;
+		if (UseRedis || UseAzureServiceBus)
 			InitialBackplaneDelay = TimeSpan.FromSeconds(5).PlusALittleBit();
 	}
 
@@ -42,8 +42,7 @@ public partial class L1L2BackplaneTests
 	private static readonly bool UseAzureServiceBus = true;
 	private static readonly string RedisConnection = "127.0.0.1:6379,ssl=False,abortConnect=false,connectTimeout=1000,syncTimeout=1000";
 
-	private static readonly string AzureServiceBusConnectionString = Environment.GetEnvironmentVariable("FUSIONCACHE_TESTS_AZURESERVICEBUS_CONNECTIONSTRING")
-		?? "Endpoint=sb://localhost;SharedAccessKeyName=RootManageSharedAccessKey;SharedAccessKey=SAS_KEY_VALUE;UseDevelopmentEmulator=true;";
+	private readonly L1L2AzureServiceBusFixture _azureServiceBusFixture;
 
 	private readonly TimeSpan InitialBackplaneDelay = TimeSpan.FromMilliseconds(300);
 	private readonly TimeSpan MultiNodeOperationsDelay = TimeSpan.FromMilliseconds(300);
@@ -54,14 +53,17 @@ public partial class L1L2BackplaneTests
 			return new RedisBackplane(new RedisBackplaneOptions { Configuration = RedisConnection }, logger: CreateXUnitLogger<RedisBackplane>());
 		if (UseAzureServiceBus)
 		{
-			var topicName = AzureServiceBusHelpers.SanitizeEntityName($"fusioncache-tests", AzureServiceBusHelpers.MaxTopicNameLength);
-			var subscriptionName = AzureServiceBusHelpers.GenerateId();
-			var adminClient = new ServiceBusAdministrationClient(AzureServiceBusConnectionString);
-			var client = new ServiceBusClient(AzureServiceBusConnectionString);
-			var clientWrapper = new AzureServiceBusClientWrapper(client, topicName, subscriptionName, CreateXUnitLogger<AzureServiceBusClientWrapper>(), new AzureServiceBusBackplaneOptions());
-			var adminWrapper = new AzureServiceBusAdminWrapper(adminClient, topicName, subscriptionName, TimeSpan.FromMinutes(10), CreateXUnitLogger<AzureServiceBusAdminWrapper>());
+			var emulator = _azureServiceBusFixture;
+			var subscriptionName = emulator.GetNextSubscriptionName();
+			var client = new ServiceBusClient(emulator.ConnectionString);
+			var options = new AzureServiceBusBackplaneOptions
+			{
+				IsAdmin = false,
+				SubscriptionName = subscriptionName,
+			};
+			var clientWrapper = new AzureServiceBusClientWrapper(client, emulator.TopicName, subscriptionName, CreateXUnitLogger<AzureServiceBusClientWrapper>(), options);
 
-			return new AzureServiceBusBackplane(clientWrapper, adminWrapper, CreateXUnitLogger<AzureServiceBusBackplane>());
+			return new AzureServiceBusBackplane(clientWrapper, NoOpAzureServiceBusAdminWrapper.Instance, CreateXUnitLogger<AzureServiceBusBackplane>());
 		}
 		return new MemoryBackplane(new MemoryBackplaneOptions() { ConnectionId = connectionId }, logger: CreateXUnitLogger<MemoryBackplane>());
 	}
