+++ tests/ZiggyCreatures.FusionCache.Tests/ZiggyCreatures.FusionCache.Tests.csproj
@@ -12,6 +12,9 @@
 		<Content Include="Snapshots\**">
 			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
 		</Content>
+		<Content Include="AzureServiceBus\L1L2BackplaneTests.servicebus-emulator.json">
+			<CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+		</Content>
 	</ItemGroup>
 
 	<ItemGroup>
@@ -22,6 +25,7 @@
 		<PackageReference Include="CacheTower" Version="0.14.0" />
 		<PackageReference Include="EasyCaching.InMemory" Version="1.9.2" />
 		<PackageReference Include="LazyCache" Version="2.4.0" />
+		<PackageReference Include="Testcontainers.ServiceBus" Version="4.13.0" />
 		<PackageReference Include="xunit.v3" Version="3.2.2" />
 		<PackageReference Include="Microsoft.Extensions.Caching.Hybrid" Version="10.5.0" />
 		<PackageReference Include="xunit.runner.visualstudio" Version="3.1.5">
